Test splitting
